fix(widget): weight-fill Android profile widget so tiles size correctly
The home-screen widget's profile tiles were sized from the launcher-reported
cell size, which differs from the real footprint by launcher and by widget
size. That caused (in turn) tiles anchored left with dead container space on
the right (ColorOS/OxygenOS on the Find X9 Pro), bottom corners clipped square,
and gap drift on resize.

Rebuild the grid as a weight-filled LinearLayout (a vertical LinearLayout of
weighted rows, each holding weighted tile cells built via RemoteViews.addView).
Weights fill the widget exactly at any size on any launcher, with uniform gaps
and rounded corners that nest in the container — no pixel-size reading at all.
bestGrid only picks the row/column split; a short last row is padded with an
invisible weighted spacer. Tile bitmaps are decoded downsampled so a widget
with many profiles can't exceed the RemoteViews bitmap budget.

Deletes ProfileTileRemoteViewsService + its GridView item layout and all the
pixel math (net simpler); folds readWidgetProfileIds + ProfileTonal into the
provider. Verified on Android 16 (API 36) and the Find X9 Pro.
