+++ android/app/src/main/kotlin/be/casperverswijvelt/sonority/ProfileWidgetProvider.kt
@@ -5,8 +5,11 @@ import android.appwidget.AppWidgetManager
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
+import android.content.res.Configuration
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import android.graphics.Color
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
 import android.widget.RemoteViews
 import es.antonborri.home_widget.HomeWidgetLaunchIntent
@@ -14,12 +17,28 @@ import es.antonborri.home_widget.HomeWidgetPlugin
 import es.antonborri.home_widget.HomeWidgetProvider
 import kotlin.math.ceil
 
-/// Home-screen widget showing a user-picked SET of profiles as a grid of square
-/// tiles, each independently tap-to-apply. Tiles are rendered to square PNGs in
-/// Flutter (keyed `tile_<profileId>`); this provider wires a GridView collection
-/// (see [ProfileTileRemoteViewsService]), picks the row/column split that makes
-/// the tiles the largest possible squares for the current widget size, and
-/// centres the grid. Per tile, a fill-in intent launches
+/// The widget's ordered profile ids: the JSON list, or the pre-multi-profile
+/// single-id key (lazy migration of widgets placed before that feature).
+fun readWidgetProfileIds(prefs: SharedPreferences, widgetId: Int): List<String> {
+    prefs.getString("profileIds_$widgetId", null)?.let { raw ->
+        return try {
+            val arr = org.json.JSONArray(raw)
+            List(arr.length()) { arr.getString(it) }
+        } catch (_: Exception) {
+            emptyList()
+        }
+    }
+    return prefs.getString("profileId_$widgetId", null)?.let { listOf(it) } ?: emptyList()
+}
+
+/// Home-screen widget showing a user-picked SET of profiles, each tile taps to
+/// apply. Tiles are rendered to PNGs in Flutter (keyed `tile_<profileId>`); this
+/// provider lays them out as a weight-filled grid — a vertical LinearLayout of
+/// weighted rows, each holding weighted tile cells (see the profile_widget_*
+/// layouts). Weights fill the widget exactly at any size, so there's no reading
+/// of the launcher-reported pixel size (which varies by launcher/size and used to
+/// cause left-anchored dead space, clipped corners, or gap drift). `bestGrid`
+/// only picks the row/column split. Per tile, a PendingIntent launches
 /// `sonority://apply?homeWidget=1&id=<profileId>` — the shape Flutter parses.
 class ProfileWidgetProvider : HomeWidgetProvider() {
     override fun onUpdate(
@@ -31,7 +50,7 @@ class ProfileWidgetProvider : HomeWidgetProvider() {
         appWidgetIds.forEach { buildWidget(context, appWidgetManager, it) }
     }
 
-    // Re-pack the grid when the user resizes the widget.
+    // Re-pack the grid when the user resizes the widget (row/column split may change).
     override fun onAppWidgetOptionsChanged(
         context: Context,
         appWidgetManager: AppWidgetManager,
@@ -43,77 +62,95 @@ class ProfileWidgetProvider : HomeWidgetProvider() {
 
     private fun buildWidget(context: Context, mgr: AppWidgetManager, id: Int) {
         val prefs = HomeWidgetPlugin.getData(context)
-        val count = maxOf(1, readWidgetProfileIds(prefs, id).size)
+        val tiles = readWidgetProfileIds(prefs, id).mapNotNull { pid ->
+            prefs.getString("tile_$pid", null)?.let { path ->
+                Tile(pid, path, prefs.getString("tileName_$pid", "") ?: "",
+                    prefs.getString("tileColor_$pid", null))
+            }
+        }
+        val views = RemoteViews(context.packageName, R.layout.profile_widget)
+        // ESSENTIAL, not redundant: on resize/update the launcher re-applies this
+        // RemoteViews onto the EXISTING view tree, where addView APPENDS. Without
+        // this clear, every update stacks another full set of tiles into the grid.
+        views.removeAllViews(R.id.grid)
 
-        // Current size in dp (portrait: MIN_WIDTH = width, MAX_HEIGHT = height).
+        if (tiles.isEmpty()) {
+            // Empty state: hide the grid, show the "Tap to pick profiles" label and
+            // make tapping it open this widget's configure screen.
+            views.setViewVisibility(R.id.grid, android.view.View.GONE)
+            views.setViewVisibility(R.id.empty, android.view.View.VISIBLE)
+            val configure = Intent(context, ProfileWidgetConfigActivity::class.java).apply {
+                action = AppWidgetManager.ACTION_APPWIDGET_CONFIGURE
+                putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, id)
+                data = Uri.fromParts("widget", id.toString(), null) // distinct per widget
+            }
+            views.setOnClickPendingIntent(
+                R.id.empty,
+                PendingIntent.getActivity(
+                    context, id, configure,
+                    PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE))
+            mgr.updateAppWidget(id, views)
+            return
+        }
+
+        views.setViewVisibility(R.id.empty, android.view.View.GONE)
+        views.setViewVisibility(R.id.grid, android.view.View.VISIBLE)
+
+        // Row/column split only — sizes come from layout weights, not these dp.
         val opts = mgr.getAppWidgetOptions(id)
         val wDp = opts.getInt(AppWidgetManager.OPTION_APPWIDGET_MIN_WIDTH).takeIf { it > 0 } ?: DEFAULT_DP
         val hDp = opts.getInt(AppWidgetManager.OPTION_APPWIDGET_MAX_HEIGHT).takeIf { it > 0 } ?: DEFAULT_DP
-        val (cols, rows) = bestGrid(count, wDp, hDp)
-        // Tiles fill their cells so the grid fills the widget with a uniform
-        // GAP_DP on every edge and between tiles (cells may be slightly
-        // non-square). The factory reads cellH to size each tile's height.
-        val cellW = maxOf(MIN_CELL_DP, (wDp - (cols + 1) * GAP_DP) / cols)
-        val cellH = maxOf(MIN_CELL_DP, (hDp - (rows + 1) * GAP_DP) / rows)
-        // Short edge drives the glyph/label sizes (see ProfileTileFactory), so the
-        // grid matches the iOS widget's spec.
-        prefs.edit()
-            .putInt("cellH_$id", cellH)
-            .putInt("tileS_$id", minOf(cellW, cellH))
-            .apply()
+        val (cols, _) = bestGrid(tiles.size, wDp, hDp)
+        val dark = (context.resources.configuration.uiMode and
+            Configuration.UI_MODE_NIGHT_MASK) == Configuration.UI_MODE_NIGHT_YES
 
-        val views = RemoteViews(context.packageName, R.layout.profile_widget)
-
-        // Collection adapter. The unique data Uri per widget id is REQUIRED —
-        // without it every placed widget shares one factory and shows the same
-        // tiles.
-        val serviceIntent = Intent(context, ProfileTileRemoteViewsService::class.java).apply {
-            putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, id)
-            data = Uri.fromParts("widget", id.toString(), null)
+        tiles.chunked(cols).forEach { rowTiles ->
+            val row = RemoteViews(context.packageName, R.layout.profile_widget_row)
+            rowTiles.forEach { row.addView(R.id.row, buildCell(context, it, dark)) }
+            // Pad a short last row so its tiles keep the full rows' width.
+            repeat(cols - rowTiles.size) {
+                row.addView(R.id.row, RemoteViews(context.packageName, R.layout.profile_widget_spacer))
+            }
+            views.addView(R.id.grid, row)
         }
-        views.setRemoteAdapter(R.id.grid, serviceIntent)
-        views.setEmptyView(R.id.grid, R.id.empty)
-        views.setInt(R.id.grid, "setNumColumns", cols)
-        val density = context.resources.displayMetrics.density
-        views.setInt(R.id.grid, "setColumnWidth", (cellW * density).toInt())
-
-        // Uniform GAP_DP between tiles on both axes (horizontal/verticalSpacing in
-        // XML) and as the outer margin (padding, px). Cells fill, so the leftover
-        // per axis works out to one GAP_DP on each edge.
-        val contentWpx = ((cols * cellW + (cols - 1) * GAP_DP) * density).toInt()
-        val contentHpx = ((rows * cellH + (rows - 1) * GAP_DP) * density).toInt()
-        val hPad = maxOf(0, ((wDp * density).toInt() - contentWpx) / 2)
-        val vPad = maxOf(0, ((hDp * density).toInt() - contentHpx) / 2)
-        views.setViewPadding(R.id.grid, hPad, vPad, hPad, vPad)
-
-        // Template must be MUTABLE and carry NO data so each item's fill-in can
-        // supply its own `id=` Uri.
-        val template = Intent(context, MainActivity::class.java).apply {
-            action = HomeWidgetLaunchIntent.HOME_WIDGET_LAUNCH_ACTION
-        }
-        var flags = PendingIntent.FLAG_UPDATE_CURRENT
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            flags = flags or PendingIntent.FLAG_MUTABLE
-        }
-        views.setPendingIntentTemplate(
-            R.id.grid, PendingIntent.getActivity(context, id, template, flags))
-
-        // The empty state says "Tap to pick profiles" — make that true: tapping
-        // it opens this widget's configure screen.
-        val configure = Intent(context, ProfileWidgetConfigActivity::class.java).apply {
-            action = AppWidgetManager.ACTION_APPWIDGET_CONFIGURE
-            putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, id)
-            data = Uri.fromParts("widget", id.toString(), null) // distinct per widget
-        }
-        views.setOnClickPendingIntent(
-            R.id.empty,
-            PendingIntent.getActivity(
-                context, id, configure,
-                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE))
-
         mgr.updateAppWidget(id, views)
-        // Force the factory to re-read (tile set / sizes may have changed).
-        mgr.notifyAppWidgetViewDataChanged(id, R.id.grid)
+    }
+
+    private fun buildCell(context: Context, tile: Tile, dark: Boolean): RemoteViews {
+        val accent = tile.colorHex?.let { runCatching { Color.parseColor(it) }.getOrNull() } ?: Color.GRAY
+        val tones = ProfileTonal.of(accent, dark)
+        return RemoteViews(context.packageName, R.layout.profile_tile_cell).apply {
+            setInt(R.id.tile_bg, "setColorFilter", tones.card)
+            setInt(R.id.widget_tile, "setColorFilter", tones.icon)
+            decodeTile(tile.path)?.let { setImageViewBitmap(R.id.widget_tile, it) }
+            setTextViewText(R.id.widget_label, tile.name)
+            setInt(R.id.widget_label, "setTextColor", tones.label)
+            // Launch MainActivity with the apply Uri Flutter parses. Distinct data
+            // per profile → distinct PendingIntent even at the same request code.
+            val intent = Intent(context, MainActivity::class.java).apply {
+                action = HomeWidgetLaunchIntent.HOME_WIDGET_LAUNCH_ACTION
+                data = Uri.parse("sonority://apply?homeWidget=1&id=${tile.id}")
+            }
+            setOnClickPendingIntent(
+                R.id.tile_cell,
+                PendingIntent.getActivity(
+                    context, tile.id.hashCode(), intent,
+                    PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE))
+        }
+    }
+
+    // Tiles are glyph PNGs Flutter renders at ~300·dpr px but shown at ~40dp.
+    // Every tile bitmap rides in the single updateAppWidget transaction, and a
+    // full-size ARGB glyph is multiple MB — decode downsampled to keep each one
+    // small so a widget with many profiles stays well within the RemoteViews
+    // bitmap budget (raises the practical ceiling; doesn't make it unbounded).
+    private fun decodeTile(path: String): Bitmap? {
+        val bounds = BitmapFactory.Options().apply { inJustDecodeBounds = true }
+        BitmapFactory.decodeFile(path, bounds)
+        if (bounds.outWidth <= 0) return null
+        var sample = 1
+        while (bounds.outWidth / (sample * 2) >= TILE_PX) sample *= 2
+        return BitmapFactory.decodeFile(path, BitmapFactory.Options().apply { inSampleSize = sample })
     }
 
     // Drop the removed widgets' per-widget prefs (the shared per-profile tiles
@@ -122,15 +159,22 @@ class ProfileWidgetProvider : HomeWidgetProvider() {
         val edit = HomeWidgetPlugin.getData(context).edit()
         appWidgetIds.forEach { id ->
             edit.remove("profileIds_$id").remove("profileId_$id")
+                // cellH_/tileS_ are no longer written; clean up any left by an
+                // older build when its widget is removed.
                 .remove("cellH_$id").remove("tileS_$id")
         }
         edit.apply()
     }
 
+    private data class Tile(
+        val id: String, val path: String, val name: String, val colorHex: String?)
+
     companion object {
-        private const val GAP_DP = 8
-        private const val MIN_CELL_DP = 24
         private const val DEFAULT_DP = 110
+
+        // Target decoded tile size (px); the glyph is shown at ~40dp, so a couple
+        // hundred px is plenty crisp while keeping bitmap memory small.
+        private const val TILE_PX = 192
 
         /// Row/column split that maximises the square tile side for [n] tiles in
         /// a [w]×[h] (dp) box. Mirrored in iOS `bestGrid`.
@@ -144,5 +188,53 @@ class ProfileWidgetProvider : HomeWidgetProvider() {
             }
             return best.first to best.second
         }
+    }
+}
+
+/// The muted-tonal treatment for an accent colour — mirrors Dart `profileTonal`
+/// so the widget matches the in-app tile and the iOS widget.
+object ProfileTonal {
+    data class Tones(val card: Int, val icon: Int, val label: Int)
+
+    fun of(accent: Int, dark: Boolean): Tones {
+        return if (dark) {
+            val card = blend(accent, 0.30f, 0xFF1B1B20.toInt())
+            Tones(card, ensureContrast(lerp(accent, Color.WHITE, 0.30f), card, Color.WHITE),
+                0xFFECECEF.toInt())
+        } else {
+            val card = blend(accent, 0.14f, 0xFFFBFBFD.toInt())
+            Tones(card, ensureContrast(accent, card, Color.BLACK), 0xFF1D1D22.toInt())
+        }
+    }
+
+    // fg at [alpha] composited over an opaque [bg].
+    private fun blend(fg: Int, alpha: Float, bg: Int) = lerp(bg, fg, alpha)
+
+    private fun lerp(a: Int, b: Int, t: Float): Int = Color.rgb(
+        (Color.red(a) + (Color.red(b) - Color.red(a)) * t).toInt().coerceIn(0, 255),
+        (Color.green(a) + (Color.green(b) - Color.green(a)) * t).toInt().coerceIn(0, 255),
+        (Color.blue(a) + (Color.blue(b) - Color.blue(a)) * t).toInt().coerceIn(0, 255))
+
+    private fun ensureContrast(fg: Int, bg: Int, toward: Int): Int {
+        var c = fg
+        var i = 0
+        while (i < 8 && contrast(c, bg) < 3.0) {
+            c = lerp(c, toward, 0.12f); i++
+        }
+        return c
+    }
+
+    private fun contrast(a: Int, b: Int): Double {
+        val la = luminance(a); val lb = luminance(b)
+        val hi = maxOf(la, lb); val lo = minOf(la, lb)
+        return (hi + 0.05) / (lo + 0.05)
+    }
+
+    private fun luminance(c: Int): Double {
+        fun ch(v: Int): Double {
+            val s = v / 255.0
+            return if (s <= 0.03928) s / 12.92 else Math.pow((s + 0.055) / 1.055, 2.4)
+        }
+        return 0.2126 * ch(Color.red(c)) + 0.7152 * ch(Color.green(c)) + 0.0722 * ch(Color.blue(c))
     }
 }
