+++ android/app/src/main/res/layout/profile_widget.xml
@@ -1,23 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!-- Neutral rounded container (adapts light/dark, fixed widget_bg_radius to match
-     the iOS widget) holding a grid of tonal profile tiles. Tiles are inset by the
-     8dp gap and use widget_tile_radius (= bg radius - 8dp) so they nest
-     concentrically inside the container. @android:id/background lets the launcher
-     auto-clip us. -->
+     the iOS widget) holding a weight-filled grid of tonal profile tiles. The grid
+     is a vertical LinearLayout the provider fills with weighted rows (see
+     ProfileWidgetProvider) — tiles split width/height by layout_weight, so they
+     fill the widget exactly at any size with no pixel-size guessing. The 4dp grid
+     padding + 4dp tile margins add up to a uniform 8dp gap (= widget_bg_radius −
+     widget_tile_radius, so tiles nest concentrically). @android:id/background lets
+     the launcher auto-clip us. -->
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@android:id/background"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/widget_container">
-    <GridView
+    <LinearLayout
         android:id="@+id/grid"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:numColumns="1"
-        android:stretchMode="none"
-        android:horizontalSpacing="8dp"
-        android:verticalSpacing="8dp"
-        android:gravity="center" />
+        android:orientation="vertical"
+        android:padding="4dp" />
     <TextView
         android:id="@+id/empty"
         android:layout_width="match_parent"
