+++ android/app/src/main/res/layout/profile_widget_spacer.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Invisible weighted filler that pads a short last row so its tiles keep the
+     same width as the full rows above (instead of stretching to fill the row). -->
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="0dp"
+    android:layout_height="match_parent"
+    android:layout_weight="1" />
